feat:update

+++ model/base.go
@@ -6,7 +6,7 @@ import (
 )
 
 type BaseModal struct {
-	Id        int32          `gorm:"primaryKey" json:"id"`
+	Id        int64          `gorm:"primaryKey" json:"id"`
 	CreatedAt time.Time      `gorm:"column:created_at" json:"created_at"`
 	UpdatedAt time.Time      `gorm:"column:updated_at" json:"updated_at"`
 	DeletedAt gorm.DeletedAt `json:"deleted_at"`
